# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicDutyOut": {
        "properties": {
          "address_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Summary"
          },
          "branch_display_name": {
            "title": "Branch Display Name",
            "type": "string"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "duty_date": {
            "title": "Duty Date",
            "type": "string"
          },
          "duty_end_at": {
            "format": "date-time",
            "title": "Duty End At",
            "type": "string"
          },
          "duty_start_at": {
            "format": "date-time",
            "title": "Duty Start At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "pharmacy_display_name": {
            "title": "Pharmacy Display Name",
            "type": "string"
          },
          "pharmacy_id": {
            "title": "Pharmacy Id",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "id",
          "city",
          "pharmacy_id",
          "pharmacy_display_name",
          "branch_display_name",
          "address_summary",
          "phone",
          "duty_date",
          "duty_start_at",
          "duty_end_at",
          "last_verified_at"
        ],
        "title": "PublicDutyOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NaviMed API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/public/pharmacies/on-duty": {
      "get": {
        "operationId": "on_duty_v1_public_pharmacies_on_duty_get",
        "parameters": [
          {
            "in": "query",
            "name": "city",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 1,
              "title": "City",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PublicDutyOut"
                  },
                  "title": "Response On Duty V1 Public Pharmacies On Duty Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "On Duty",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/public/pharmacies/{pharmacy_id}": {
      "get": {
        "operationId": "pharmacy_public_v1_public_pharmacies__pharmacy_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pharmacy_id",
            "required": true,
            "schema": {
              "title": "Pharmacy Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PublicDutyOut"
                  },
                  "title": "Response Pharmacy Public V1 Public Pharmacies  Pharmacy Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pharmacy Public",
        "tags": [
          "public"
        ]
      }
    }
  }
}